+Security, +Package Manager, +User Integration
Implemented yarn to manage dependencies. Created a migration for a local-db oriented user-model. Added the model to the sequelize definitions. Created a registration-route, which references a user-model markup file via swagger to define request body parameters. Added a global api-key tied to the Global-Security handler to be required when making requests.