Major terminology Refactor. Static hard-coded classes to Resources folder
authorization on request-object now authentication. all error classes relegated to a 'resources' folder, and they're hard coded classes with defined properties. Ability to override the token-key with an api-key which will bypass a check for a token and default the authentication header data to feature uid: 0. Fixed errorReducer to ALWAYS return ServerError typed classes, logging along the way. (Log inside of the reducer instead of when we process the reduced error afterwards). App.logger.error wrapped in a function to typecheck for ServerErrors, and will log them accordingly to the mongodb database.